# openapi/brokkr-v1.json
{
  "components": {
    "schemas": {
...
      "AgentEvent": {
        "description": "Represents an agent event in the database.",
        "example": {
          "agent_id": "123e4567-e89b-12d3-a456-426614174001",
          "created_at": "2023-01-01T00:00:00Z",
          "deleted_at": null,
          "deployment_object_id": "123e4567-e89b-12d3-a456-426614174002",
          "event_type": "DEPLOYMENT",
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "message": "Deployment completed successfully",
          "status": "SUCCESS",
          "updated_at": "2023-01-01T00:00:00Z"
        },
        "properties": {
          "agent_id": {
            "description": "ID of the agent associated with this event.",
            "example": "123e4567-e89b-12d3-a456-426614174001",
            "format": "uuid",
            "type": "string"
          },
          "created_at": {
            "description": "Timestamp when the event was created.",
            "example": "2023-01-01T00:00:00Z",
            "format": "date-time",
            "type": "string"
          },
          "deleted_at": {
            "description": "Timestamp for soft deletion, if applicable.",
            "example": "null",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "deployment_object_id": {
            "description": "ID of the deployment object associated with this event.",
            "example": "123e4567-e89b-12d3-a456-426614174002",
            "format": "uuid",
            "type": "string"
          },
          "event_type": {
            "description": "Type of the event.",
            "example": "DEPLOYMENT",
            "type": "string"
          },
          "id": {
            "description": "Unique identifier for the event.",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid",
            "type": "string"
          },
          "message": {
            "description": "Optional message providing additional details about the event.",
            "example": "Deployment completed successfully",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "description": "Status of the event (e.g., \"SUCCESS\", \"FAILURE\", \"IN_PROGRESS\", \"PENDING\").",
            "example": "SUCCESS",
            "type": "string"
          },
          "updated_at": {
            "description": "Timestamp when the event was last updated.",
            "example": "2023-01-01T00:00:00Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "agent_id",
          "deployment_object_id",
          "event_type",
          "status"
        ],
        "type": "object"
      },
      "AgentFleetStatusResponse": {
        "description": "Response body for the per-agent fleet-status detail view: the agent's fleet\nrecord plus its most recent events (newest first).",
        "properties": {
          "recent_events": {
            "description": "The agent's most recent events, newest first (up to 20).",
            "items": {
              "$ref": "#/components/schemas/AgentEvent"
            },
            "type": "array"
          },
          "record": {
            "$ref": "#/components/schemas/FleetAgentRecord",
            "description": "The per-agent fleet record (same shape as the rollup entries)."
          }
        },
        "required": [
          "record",
          "recent_events"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Wire format for every 4xx/5xx response body in the v1 API.",
        "properties": {
          "code": {
            "description": "Stable, machine-readable error code (e.g. `agent_not_found`).",
            "type": "string"
          },
          "details": {
            "additionalProperties": {},
            "description": "Optional structured context. Keys and value shapes are documented per\nerror code when present.",
            "nullable": true,
            "type": "object"
          },
          "message": {
            "description": "Human-readable message; not stable and may change between versions.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "FleetAgentRecord": {
        "description": "A per-agent fleet record: measured signals only, no health verdicts.\n\nAll time-relative fields (`heartbeat_age_seconds`, `seconds_since_last_event`)\nare computed on read as `now - timestamp`, clamped to be non-negative.",
        "properties": {
          "agent_id": {
            "description": "The agent's unique identifier.",
            "format": "uuid",
            "type": "string"
          },
          "claimed_work_orders": {
            "description": "Number of work orders currently CLAIMED by this agent.",
            "format": "int64",
            "type": "integer"
          },
          "connected_since": {
            "description": "When the current WebSocket connection was established, if connected.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "health_degraded": {
            "description": "Count of this agent's deployment-health records with status `degraded`.",
            "format": "int64",
            "type": "integer"
          },
          "health_failing": {
            "description": "Count of this agent's deployment-health records with status `failing`.",
            "format": "int64",
            "type": "integer"
          },
          "heartbeat_age_seconds": {
            "description": "Seconds since the last heartbeat (`now - last_heartbeat`, clamped >= 0).",
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "last_event_at": {
            "description": "Timestamp of this agent's most recent (non-deleted) event, if any.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "last_heartbeat": {
            "description": "The agent's last recorded heartbeat timestamp.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "description": "The agent's name.",
            "type": "string"
          },
          "pending_object_count": {
            "description": "Number of pending (not-yet-acknowledged) deployment objects targeted at\nthis agent.",
            "format": "int64",
            "type": "integer"
          },
          "pending_work_orders": {
            "description": "Number of PENDING work orders this agent is eligible to claim.",
            "format": "int64",
            "type": "integer"
          },
          "seconds_since_last_event": {
            "description": "Seconds since the last event (`now - last_event_at`, clamped >= 0).",
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "status": {
            "description": "The agent's lifecycle status (e.g. \"ACTIVE\").",
            "type": "string"
          },
          "ws_connected": {
            "description": "Whether the agent currently holds a broker↔agent WebSocket connection.",
            "type": "boolean"
          }
        },
        "required": [
          "agent_id",
          "name",
          "status",
          "ws_connected",
          "pending_object_count",
          "pending_work_orders",
          "claimed_work_orders",
          "health_failing",
          "health_degraded"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "admin_pak": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "agent_pak": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "generator_pak": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Brokkr broker: the central control-plane API server for distributing Kubernetes objects across clusters.",
    "license": {
      "name": "Elastic-2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    },
    "title": "brokkr-broker",
    "version": "0.7.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/agents/{id}/fleet-status": {
      "get": {
        "operationId": "get_agent_fleet_status",
        "parameters": [
          {
            "description": "ID of the agent",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentFleetStatusResponse"
                }
              }
            },
            "description": "Fleet record plus recent events for the agent"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          }
        ],
        "tags": [
          "fleet"
        ]
      }
    },
...
    "/fleet": {
      "get": {
        "operationId": "list_fleet",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FleetAgentRecord"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Per-agent fleet records (measured values only)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "admin_pak": []
          }
        ],
        "tags": [
          "fleet"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}